Move StateMachine to ds header
